Transfer tests
of test_frame.py
to test_frame_color.py, test_frame_groupby.py and test_frame_subplots.py
